+++ Pinta/Actions/Edit/PreferencesAction.cs
@@ -47,6 +47,7 @@ internal sealed class PreferencesAction : IActionHandler
 			defaultCanvasSurroundColor,
 			settings.GetSetting (SettingNames.PASTE_EXTERNAL_IMAGES_TO_NEW_LAYER, false),
 			extendedPaletteRows,
+			PintaCore.Palette.CurrentPalette.Colors.Count,
 			toolboxClassicLayout,
 			toolSettingsWrapRows,
 			(PopoverHintMode) settings.GetSetting (SettingNames.POPOVER_HINT_MODE, (int) PopoverHintMode.All),
@@ -86,6 +87,9 @@ internal sealed class PreferencesAction : IActionHandler
 				settings.PutSetting (SettingNames.EXTENDED_PALETTE_ROWS, dialog.ExtendedPaletteRows);
 				PintaCore.Palette.CurrentPalette.LoadDefault (dialog.ExtendedPaletteRows);
 			}
+
+			if (dialog.PaletteSize != PintaCore.Palette.CurrentPalette.Colors.Count)
+				PintaCore.Palette.CurrentPalette.Resize (dialog.PaletteSize);
 
 			foreach (Document document in PintaCore.Workspace.OpenDocuments)
 				((CanvasWindow) document.Workspace.CanvasWindow).CanvasSurroundColor = selectedCanvasSurroundColor;